+++ src/main/java/org/filespace/model/intermediate/UserInfo.java
@@ -1,6 +1,6 @@
 package org.filespace.model.intermediate;
 
 public interface UserInfo {
-    public Long getId();
+    public Integer getId();
     public String getUsername();
 }
