+++ src/main/java/org/filespace/repositories/FilespaceRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FilespaceRepository extends JpaRepository<Filespace, Long> {
+public interface FilespaceRepository extends JpaRepository<Filespace, Integer> {
 
 }
